--- a/mani_skill/agents/controllers/__init__.py
+++ b/mani_skill/agents/controllers/__init__.py
@@ -10,7 +10,7 @@ from .pd_ee_pose import (
     PDEEPosControllerConfig,
     PDEEPoseController,
     PDEEPoseControllerConfig,
-    PDEEPoseControllerConfig_NEW
+    PDEEPoseControllerConfig_Fast
 )
 from .pd_joint_vel import PDJointVelController, PDJointVelControllerConfig
 from .pd_joint_pos_vel import PDJointPosVelController, PDJointPosVelControllerConfig