@StoneT2000 2025-08-01
we are moving to try and make a new package release soon so I will just merge this to a dev branch and make the final edits for some code consistency and what note.

Only change I will make is likely I will remove the new config and instead we keep the same config but with an option for choice of solver, default being the LM based one.

The new solver code will also go into the kinematics.py file as well.